insert hints on every decision
the plugin now works a bit differently, we insert a card for every decision and
have the card decide whether it should show itself based on the available
snippets.

This seems like a more consistent way of handling the hints, if a new snippet
arrives we already have our hints ready.